Update docker-compose and pre-commit configurations
The LDAP password in the Semaphore docker-compose file has been updated to pull from an environment variable instead of being hardcoded. The pre-commit configuration has been revised, updated to use the latest versions of the pre-commit-hooks and gitleaks repositories, as well as adding the 'checkov' tool from Bridgecrew for improved checking of yaml files.